tmux: exit copy-mode on mouse up, not down
This way, if I'm scrolled back, searching in copy mode, I can use the
mouse to select some found text. Exiting copy mode on mouse-down was
causing copy mode to exit as soon as I tried to use the mouse to select.
Exiting on mouse up instead allows you to click and drag without
exiting, but if you do a normal click and release it exits as expected.