fix this shit
Trying to get tmux to exit copy-mode when clicked, but not when
double-clicked, or at the end of a click-and-drag, this seems to be an
intractable problem. Binding cancel to mouse-up works on one machine and
not others. All in all, it's not important enough to fiddle with any
longer.

Totally unrelated is the search binding in this commit -- that's just a
binding I meant to add earlier.